Fixed issues with cropped brush
By pre cropping the source image, the issues of brushes that had a cropping viewbox are solved. I tried and tried and tried to solve this through getting the matrix right but after 10 hours of trial and error, google and too much coffee I decided to just chop the image, before the rest happens.

I think it's okay.

I'm not totally sure if a brush transform might get upset by this, if anything I think it would make them better too but I need to test.

I'll add an xps now to show what this fixed.